Move unused labmda
