+++ backend/src/lambda_notused/accountsResolver/helpers/dynamoDbCommand.ts
