+++ backend/src/lambda_notused/accountsResolver/helpers/eventBridge.ts
